pland/backend/requirements.txt
# This file is autogenerated by pip-compile with Python 3.11
# To update, run:
#
#    pip-compile requirements.in
#
aiofiles==23.2.1
    # via -r requirements.in
fastapi==0.104.1
    # via -r requirements.in
httpx==0.25.2
    # via -r requirements.in
jinja2==3.1.2
    # via -r requirements.in
numpy==1.24.3
    # via -r requirements.in
onnx==1.15.0
    # via -r requirements.in
onnxruntime==1.16.3
    # via -r requirements.in
opencv-python-headless==4.8.1.78
    # via -r requirements.in
pandas==2.0.3
    # via -r requirements.in
pillow==10.0.1
    # via -r requirements.in
pydantic==2.5.0
    # via -r requirements.in
pydantic-settings==2.1.0
    # via -r requirements.in
pytest==7.4.3
    # via -r requirements.in
pytest-asyncio==0.21.1
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
python-multipart==0.0.6
    # via -r requirements.in
rembg==2.0.50
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
scikit-image==0.21.0
    # via -r requirements.in
scikit-learn==1.3.0
    # via -r requirements.in
timm==0.9.12
    # via -r requirements.in
torch==2.1.0
    # via -r requirements.in
torchvision==0.16.0
    # via -r requirements.in
tqdm==4.66.1
    # via -r requirements.in
uvicorn[standard]==0.24.0
    # via -r requirements.in
